Add an option to limit DPL concurrency with pool counter
The previous patch tried to eliminate the common case of
duplicated work by DPL from parsing pages all using the
same template, causing the same query to pile-up.

This patch aims to be more of an emergency break. It allows
limiting the number of concurrent DPL queries to be given from
a wiki at one time. The idea is that this would prevent one
wiki from using more than its fair share of server resources.

This limit is per-wiki, not per-page or per-query. If the limit
is reached, an error message is output saying the servers are
overloaded. The error can be disabled by doing suppresserrors=all
(It seems like people might still want this error, even if they
are suppressing other errors, hence the new option).

This feature is disabled by default unless additional config added.
Recommended config:
$wgPoolCounterConf = [ 'DPL' => [
      'class' => PoolCounter_Client::class,
      'timeout' => 0,
      'workers' => 15,
      'maxqueue' => 15,
 ] ];
If you want to be more paranoid, or the DB seems under heavy load,
use 5 instead of 15 for workers and maxqueue.

Bug: T262391
Bug: T262240
Bug: T263220
Change-Id: If7e6c0fe9c3ab364312e2c53c48e1662b5a75844